Task2

+++ Tyuiu.GalyameevMR.Sprint5.Task2.V9.Lib/DataService.cs
@@ -12,48 +12,39 @@ namespace Tyuiu.GalyameevMR.Sprint5.Task2.V9.Lib
     {
         public string SaveToFileTextData(int[,] matrix) 
         {
-            string path = $@"{Directory.GetCurrentDirectory()}\OutPutFileTask2.txt";
-            FileInfo fileInfo = new FileInfo(path);
-            bool fileExists = fileInfo.Exists;
+            string path = $@"{Directory.GetCurrentDirectory()}\OutPutFileTask2.csv";
+
+            FileInfo fileinfo = new FileInfo(path);
+            bool fileex = fileinfo.Exists;
+
+            if (fileex) File.Delete(path);
+
             int rows = matrix.GetUpperBound(0) + 1;
             int columns = matrix.Length / rows;
-            if (fileExists)
-            {
-                File.Delete(path);
-            };
+
             for (int i = 0; i < rows; i++)
             {
                 for (int j = 0; j < columns; j++)
                 {
-                    if (matrix[i,j] % 2 != 0)
-                    {
-                        matrix[i, j] = 0;
-                    }
+                    if (matrix[i, j] % 2 != 0) matrix[i, j] = 0;
+
                 }
             }
+
             string str = "";
+
             for (int i = 0; i < rows; i++)
             {
                 for (int j = 0; j < columns; j++)
                 {
-                    if (j != columns - 1)
-                    {
-                        str = str + matrix[i, j] + ";";
-                    }
-                    else
-                    {
-                        str = str + matrix[i, j];
-                    }
+                    if (j != columns - 1) str = str + matrix[i, j] + ";";
+                    else str = str + matrix[i, j];
                 }
-                if (i != rows - 1)
-                {
-                    File.AppendAllText(path, str + Environment.NewLine);
-                }
-                else
-                {
-                    File.AppendAllText(path, str);
-                }
+                if (i != rows - 1) File.AppendAllText(path, str + Environment.NewLine);
+                else File.AppendAllText(path, str);
+                str = "";
             }
+
             return path;
         }
     }
